Coerce curl_escape/curl_unescape arguments to character.
This copies the default behavior of `RCurl`, which automatically coerces the
`url` argument to a character. This has the side-effect of fixing `NULL`
handling (see hadley/bigrquery#54 and hadley/httr#259).

If the goal was to explicitly *avoid* coercion (since the functions are
documented as consuming a character vector), I can reduce this to simply do
better `NULL` handling.